made doubles out of some floats

+++ RDynState5NAsigmaseason6Age/src/RDynState.cpp
@@ -30,8 +30,8 @@ typedef unsigned UFINT;
 typedef float  (*FTYPE)[SPP2CAPACITY]; //matalloc(sizeof(float), (void *)0, 2, kSpp1Cap, kSpp2Cap); /* make pointer called FTYPE for FF0Star and FF1 arrays (that only has states and stores V for best choices) */
 typedef float  (*FCTYPE)[SPP2CAPACITY][3]; // (FCTYPE) matalloc(sizeof(float), (void*) 0,3, kSpp1Capacity, kSpp2Capacity,kNPatch) ;  /* make pointer called FCTYPE for FF0 array (that has states and patch, storing V for all patches/choices)_  */
 typedef float  (*ITYPE)[SPP1CAPACITY][SPP2CAPACITY][3]; //  matalloc(sizeof(float), (void *)0, 4, MAXHORIZON, kSpp1Capacity,kSpp2Capacity,kNPatch);  /* make pointer called ITYPE optimal choice array*/
-typedef float  (*PTYPE)[MAXHORIZON][NOSPEC][NOSIZES][MAXNOINC]; /* make pointer called PTYPE for information on statistical distribution prey*/
-typedef float  (*ATYPE)[MAXHORIZON][NOSPEC][NOSIZES*MAXNOINC - 1];  /* make pointer called ATYPE for information on statistical distribution prey aggregated over sizes*/
+typedef double (*PTYPE)[MAXHORIZON][NOSPEC][NOSIZES][MAXNOINC]; /* make pointer called PTYPE for information on statistical distribution prey*/
+typedef double (*ATYPE)[MAXHORIZON][NOSPEC][(NOSIZES*MAXNOINC) - 1];  /* make pointer called ATYPE for information on statistical distribution prey aggregated over sizes*/
 typedef float  (*PITYPE)[NOSPEC][NOSIZES];                         // matalloc(sizeof(float), (void *)0, 3, MAXHORIZON, NOSPEC, NOSIZES);; /* make pointer called PITYPE for price information*/
 
 
@@ -579,7 +579,7 @@ Rprintf("Start of DynStateF\n");
 	        for (inc4 = 0; inc4 < noInc; inc4++){
 	          for (inc5 = 0; inc5 < noInc; inc5++){
 	            
-		theLndParmsAgg[i][t][s][inc0 + inc1 + inc2 + inc3 + inc4 + inc5] +=  theLndParms[i][t][s][0][inc0] * theLndParms[i][t][s][1][inc1] *  theLndParms[i][t][s][2][inc2]  * theLndParms[i][t][s][3][inc3] * theLndParms[i][t][s][4][inc4] *theLndParms[i][t][s][5][inc5] ;
+		theLndParmsAgg[i][t][s][inc0 + inc1 + inc2 + inc3 + inc4 + inc5] +=  theLndParms[i][t][s][0][inc0] * theLndParms[i][t][s][1][inc1] *  theLndParms[i][t][s][2][inc2]  * theLndParms[i][t][s][3][inc3] * theLndParms[i][t][s][4][inc4] * theLndParms[i][t][s][5][inc5] ;
 	      }
 	    }
 	  }
